Add files via upload
Instead of an API it was used a local data file. This would simulate the JSON object from an API (to see how I would make a backend request, please check my tours project). In this project they key aspect was the Set object which I used in order to make the buttons work (for unique values)